refactor: extract lifespan module
